--- front/src/components/molecules/clickable/BasicButton/BasicButton.jsx
@@ -1,16 +1,0 @@
-import React from 'react'
-import BasicButtonStyle from './BasicButton.style'
-const BasicButtons = () => {
-  return (
-    <div>
-      <BasicButtonStyle>
-       Se connecter
-      </BasicButtonStyle>
-      <BasicButtonStyle primary>
-       Se connecter
-      </BasicButtonStyle>
-    </div>
-  )
-}
-
-export default BasicButtons
