+++ front/src/components/molecules/clickable/LoginButtons/LoginButton.jsx
@@ -1,0 +1,15 @@
+import React from 'react'
+import LoginButtonStyle from "./LoginButtonStyle.style"
+
+const LoginButtons = () => {
+  return (  
+      <LoginButtonStyle>
+        <svg width="21" height="21" viewBox="0 0 37 40" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <path fill-rule="evenodd" clip-rule="evenodd" d="M23.6882 5.02441C31.3848 7.14999 36.8816 14.0101 36.8816 22.1259C36.8816 31.9809 28.6855 40.0002 18.4907 40.0002C8.29582 40.0002 0 31.9809 0 22.1259C0 14.0101 5.59729 7.15013 13.2931 5.02441V10.1452C8.3955 12.0777 4.99729 16.7152 4.99729 22.1259C4.99729 29.3724 10.9943 35.1695 18.4907 35.1695C25.8872 35.1695 31.984 29.3724 31.984 22.1259C31.984 16.7152 28.4857 12.0774 23.6882 10.1452V5.02441Z" fill="#F3EDD7"/>
+        <path fill-rule="evenodd" clip-rule="evenodd" d="M18.4912 0C20.0902 0 21.4895 1.35272 21.4895 2.89843V19.1306C21.4895 20.7732 20.0902 22.029 18.4912 22.029C16.7919 22.029 15.3926 20.7729 15.3926 19.1306V2.89843C15.3926 1.35264 16.7919 0 18.4912 0V0Z" fill="#F3EDD7"/>
+        </svg>
+      </LoginButtonStyle>
+  )
+}
+
+export default LoginButtons
