+++ front/src/components/molecules/clickable/LoginButtons/LoginButtonStyle.style.js
@@ -1,13 +1,13 @@
 import styled from 'styled-components';
 
-const LoginButtonsStyleLeft = styled.button`
+const LoginButtonStyle = styled.button`
     background-color: #a44d4d;
     padding: 0.35em 1em;
     border: none;
-    border-radius: 14px 0 0 14px;
+    border-radius: 0 14px 14px 0;
         &:hover {
             background-color: #ce6060;
         }
 `;
 
-export default LoginButtonsStyleLeft;
+export default LoginButtonStyle;
