+++ front/src/components/molecules/clickable/LoginButtons/Separator.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import SeparatorStyle from "./SeparatorStyle.style"
+
+const Separator= () => {
+  return (
+      <SeparatorStyle>
+        <svg width="15" height="25" viewBox="0 0 3 40" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <rect width="4" height="54" fill="#F3EDD7"/>
+        </svg>
+      </SeparatorStyle>
+  )
+}
+
+export default Separator
