+++ front/src/components/molecules/clickable/LoginButtons/ThemeButton.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import ThemeButtonStyle from "./ThemeButtonStyle.style"
+
+const ThemeButtons = () => {
+  return (
+      <ThemeButtonStyle>
+        <svg width="21" height="21" viewBox="0 0 34 35" fill="none" xmlns="http://www.w3.org/2000/svg">
+        <path d="M33.4696 25.7246C30.0793 26.8335 26.4315 26.8831 23.0123 25.8666C19.5931 24.8501 16.5648 22.8159 14.3309 20.0349C12.0969 17.2539 10.7634 13.8583 10.5081 10.3003C10.2527 6.74237 11.0875 3.19107 12.9014 0.119562C10.2024 1.0014 7.75763 2.52484 5.77681 4.55919C3.796 6.59354 2.33828 9.07806 1.52871 11.7996C0.719145 14.5212 0.581903 17.3985 1.12876 20.1847C1.67562 22.971 2.89025 25.583 4.66844 27.7966C6.44664 30.0102 8.73532 31.7595 11.3381 32.8942C13.9409 34.0289 16.7802 34.5153 19.6123 34.3115C22.4443 34.1078 25.1847 33.22 27.5983 31.7244C30.0119 30.2288 32.0266 28.1699 33.4696 25.7246Z" fill="#F3EDD7"/>
+        </svg>
+      </ThemeButtonStyle>
+  )
+}
+
+export default ThemeButtons
