+++ front/src/components/molecules/clickable/LoginButtons/ThemeButtonStyle.style.js
@@ -1,13 +1,13 @@
 import styled from 'styled-components';
 
-const LoginButtonsStyleRight = styled.button`
+const ThemeButtonStyle = styled.button`
     background-color: #a44d4d;
     padding: 0.35em 1em;
     border: none;
-    border-radius: 0 14px 14px 0;
+    border-radius: 14px 0 0 14px;
         &:hover {
             background-color: #ce6060;
         }
 `;
 
-export default LoginButtonsStyleRight;
+export default ThemeButtonStyle;
